requirements.txt
Flask==1.1.2
mysql-connector-python==8.0.23
Werkzeug==1.0.1
python-dotenv==0.17.1
boto3==1.17.93